Moved api/cli to cli and api/rpc to rpc
