+++ rpc/handler_test.go
